+++ Documentation/changelog.md
@@ -4,5 +4,11 @@ Version control history:
 * Version 1.0.2-1 271117
 	* First version release
 	
+** Version 1.1-3 051217
+	* logging and enhanced error handling setup.
+	* resize screen and layout.
+	* some code optimisations..
+	* Optimise text formatting in favourites screen. 
+	* Ability to view and access favourites database from show TV details screens.